RepositoryCache: limit cache size
Unbounded cache size is inacceptable.

I don't see why a full-fledged repository needs to be used here, either,
since this cache requires none of the consistency or durability guarantees
made by it (and bought with a performance impact).

This implementation uses an LRUCache to effectively bound the size,
checks for ENOSPC when writing to the cache and uses tempfiles
with no flushes. That way most writes will never hit the disk
(if there is one at /tmp), but stay in the page cache, which is the
intention anyway.

If a write results in ENOSPC, then exponential backoff is invoked.